comments
Added a few points where i saw necessary, overview:
Make sure that the bot accounts for most use cases, people will try their best (knowingly or unknowingly) to break the bot and we want it to work 99% of the time. 
You can respond to the /command that was run with await ctx.respond(***) instead of ctx.send(***) this prevents that error message from popping up saying that the bot didn't respond. 
You can respond with embeds, which imo look nicer, look at the help command to see the syntax for this, i didn't want to suggest changing all of them if you didn't like them.
Limbo should be able to remove users from the queue if they're AFK or for whatever reason. (Add /remove-user command that only he can use)
People should be able to view the queue (/view-queue)
people should also be able to remove themselves from the queue. (/leave) 

It looks good so far and ill add these later when i get home